lib3h: Use PathBuf for work_dir field
This came up while I was working on holochain-rust#660. The fact that
it used String before meant that String based function signatures
were scattered around the code surrounding RealEngineConfig.

This is a breaking change, but it is easy to address (often as simple as changing
String::from(_) to PathBuf::from(_))